Add VideoFrameDecoder.
